VisuAlphabet/requirements.txt
torch~=2.0.1
diffusers~=0.21.4
transformers~=4.34.0
flask~=3.0.0
Pillow~=10.0.1